# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for SMM-auto application",
        "title": "SMM-Auto API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/payments/paymentMethod/add": {
            "post": {
                "description": "Add payment method (card)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Add payment method",
                "operationId": "AddPaymentMethod",
                "parameters": [
                    {
                        "description": "Card data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.AddPaymentMethodRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns array of subscription plans (subscription type can be checked here : https://stripe.com/docs/api/plans)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error identifying user",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "417": {
                        "description": "Error identifying user",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
        "/payments/plans": {
            "get": {
                "description": "Get all available subscription plans",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get subscription plans",
                "operationId": "GetSubPlans",
                "responses": {
                    "200": {
                        "description": "Returns array of subscription plans (subscription type can be checked here : https://stripe.com/docs/api/plans)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
        "/payments/subscription": {
            "post": {
                "description": "Create subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create subscription",
                "operationId": "CreateSub",
                "parameters": [
                    {
                        "description": "Card data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error identifying user",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "417": {
                        "description": "Error identifying user",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.AddPaymentMethodRequest": {
            "type": "object",
            "properties": {
                "cardNumber": {
                    "type": "string"
                },
                "cvc": {
                    "type": "string"
                },
                "expMonth": {
                    "type": "integer"
                },
                "expYear": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.CreateSubscriptionRequest": {
            "type": "object",
            "properties": {
                "subscriptionType": {
                    "type": "string"
                }
            }
        },
...
        "jsonHelper.ApiError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}